requirements.txt
streamlit
google-generativeai
# openai (can be removed if you fully switch)